+++ backend/src/controllers/reports/Tab_14_rpt_P_BudgetSummary.js
@@ -7,7 +7,7 @@ const controller = useController(model, what);
 
 // Template and data reading
 const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
+const { getReport, getDocumentApiBody, pdfConfig, getCurrentDate } = utils;
 controller.getReport = getReport;
 
 /**
@@ -21,7 +21,6 @@ controller.Tab_14_rpt_P_BudgetSummary = async (request, reply) => {
   controller.userRequires(request, "PMO-Reports-Capability", reply);
   try {
     // Get the data from the database.
-    const getDate = async () => new Date();
     const projectId = Number(request.query.project);
     const result = {
       project: await model.findById(projectId),
@@ -31,7 +30,7 @@ controller.Tab_14_rpt_P_BudgetSummary = async (request, reply) => {
       change_request: await model.getChangeRequests(projectId),
       contracts: await model.getContracts(projectId),
       contract_summaries: await model.getContractSummary(projectId),
-      report_date: await getDate(),
+      report_date: await getCurrentDate(),
     };
 
     const body = await getDocumentApiBody(result, "Tab_14_rpt_P_BudgetSummary.docx");