+++ backend/src/controllers/reports/Tab_17_rpt_P_Status_MostRecent.js
@@ -7,7 +7,7 @@ const controller = useController(model, what);
 
 // Template and data reading
 const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
+const { getReport, getDocumentApiBody, pdfConfig, getCurrentDate } = utils;
 controller.getReport = getReport;
 
 /**
@@ -21,7 +21,6 @@ controller.Tab_17_rpt_P_Status_MostRecent = async (request, reply) => {
   controller.userRequires(request, "PMO-Reports-Capability", reply);
   try {
     // Get the data from the database.
-    const getDate = async () => new Date();
     const projectId = Number(request.query.project);
     const result = {
       project: await model.findById(projectId),
@@ -29,7 +28,7 @@ controller.Tab_17_rpt_P_Status_MostRecent = async (request, reply) => {
       milestones: await model.getMilestones(projectId),
       alignment: await model.getStrategicAlignment(projectId),
       status: await model.findMostRecentStatusById(projectId),
-      report_date: await getDate(),
+      report_date: await getCurrentDate(),
     };
 
     const body = await getDocumentApiBody(result, "Tab_17_rpt_P_Status_MostRecent.docx");