+++ backend/src/controllers/reports/Tab_18_rpt_P_StatusSummary.js
@@ -7,7 +7,7 @@ const controller = useController(model, what);
 
 // Template and data reading
 const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
+const { getReport, getDocumentApiBody, pdfConfig, getCurrentDate } = utils;
 controller.getReport = getReport;
 
 /**
@@ -21,7 +21,6 @@ controller.Tab_18_rpt_P_StatusSummary = async (request, reply) => {
   controller.userRequires(request, "PMO-Reports-Capability", reply);
   try {
     // Get the data from the database.
-    const getDate = async () => new Date();
 
     const projectId = Number(request.query.project);
     // Get the data from the database.
@@ -32,7 +31,7 @@ controller.Tab_18_rpt_P_StatusSummary = async (request, reply) => {
       alignment: await model.getStrategicAlignment(projectId),
       statuses: await model.getProjectStatuses(projectId),
       lessons: await model.getLessonsLearned(projectId),
-      report_date: await getDate(),
+      report_date: await getCurrentDate(),
     };
     const body = await getDocumentApiBody(result, "Tab_18_rpt_P_StatusSummary.docx");
     const pdf = await cdogs.api.post("/template/render", body, pdfConfig);