+++ backend/src/controllers/reports/Tab_25_rpt_PA_LessonsLearnedbyCategory.js
@@ -7,7 +7,7 @@ const controller = useController(model, what);
 
 // Template and data reading
 const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
+const { getReport, getDocumentApiBody, pdfConfig, getCurrentDate } = utils;
 controller.getReport = getReport;
 
 /**
@@ -23,10 +23,9 @@ controller.Tab_25_rpt_PA_LessonsLearnedbyCategory = async (request, reply) => {
   controller.userRequires(request, "PMO-Reports-Capability", reply);
   try {
     // Get the data from the database.
-    const getDate = async () => new Date();
 
     const result = {
-      report_date: await getDate(),
+      report_date: await getCurrentDate(),
       lessonsLearnedByCategory,
     };